+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
 
+## [8.2.0][] - 2025-12-06
+
 ### Updated
 
 - Adding minified and unminified UMD versions in package distribution
@@ -222,7 +224,8 @@ And pass `featureName` prop with `!` prefix to keep the same behavior on WebApp
 [4.0.5]: https://github.com/willmendesneto/reactor-feature-toggle/tree/v4.0.5
 
 
-[Unreleased]: https://github.com/willmendesneto/reactor-feature-toggle/compare/v8.1.0...HEAD
+[Unreleased]: https://github.com/willmendesneto/reactor-feature-toggle/compare/v8.2.0...HEAD
+[8.2.0]: https://github.com/willmendesneto/reactor-feature-toggle/compare/v8.1.0...v8.2.0
 [8.1.0]: https://github.com/willmendesneto/reactor-feature-toggle/compare/v8.0.0...v8.1.0
 [8.0.0]: https://github.com/willmendesneto/reactor-feature-toggle/compare/v7.0.0...v8.0.0
 [7.0.0]: https://github.com/willmendesneto/reactor-feature-toggle/compare/v6.1.0...v7.0.0